stockx-discord/requirements.txt
cloudscraper=1.2.71
requests=^2.28.2
discord.py=^2.3.1